Ensure pg is stopped cleanly before `compile` exits
The platform does send a TERM when `compile` exits, but it doesn't wait for the process to stop and doesn't handle any cases where the TERM doesn't work. This could result in postgres being in an odd state when the background script is executed later. Instead use `pg_ctl` with `-w` option to ensure pg is stopped.